refactor: account service
